+++ MovingSprite.cpp
@@ -2,7 +2,7 @@
 #include <SDL.h>
 
 namespace ge{
-	void MovingSprite::translate(int x, int y)
+	void MovingSprite::translate(float x, float y)
 	{				
 		position.setX(x);
 		position.setY(y);
