+++ MovingSprite.h
@@ -6,10 +6,10 @@ namespace ge {
 	class MovingSprite : public Sprite
 	{
 	public:		
-		void translate(int x, int y);
+		void translate(float x, float y);
 	protected:
 		Vector2 velocity;
-		MovingSprite(int posX, int posY, int width, int height, const char* imgDestination) : Sprite(posX, posY, width, height, imgDestination) {};
+		MovingSprite(float posX, float posY, int width, int height, const char* imgDestination) : Sprite(posX, posY, width, height, imgDestination) {};
 	private:		
 		MovingSprite(const MovingSprite&) = delete;
 		const MovingSprite& operator=(const MovingSprite&) = delete;
